refactor(about): extract career history into CareerSection
- Move Work career data, types, photo-grid helpers and video mockups
  out of the about page into a reusable components/CareerSection.tsx
- About page now renders <CareerSection label="Work" />
- No visual change
